Layout of the clubs list works
TODO: populate the cardview with data and clean the code

+++ app/src/main/java/bum/icehockeyfordummies/database/ClubDAO.java
@@ -30,7 +30,7 @@ public interface ClubDAO {
 
     // INSERT, UPDATE and DELETE queries
     @Insert
-    int insert(ClubEntity club);
+    long insert(ClubEntity club);
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     void insertAll(List<ClubEntity> clubs);
